+++ tests/disable_test_plugin_python_parall.py
@@ -56,7 +56,7 @@ class TestPluginPythonParall(unittest.TestCase):
         Test pandora + plugin_libsgm
 
         """
-        user_cfg = pandora.read_config_file('conf/sgm_python_parall.json')
+        user_cfg = pandora.read_config_file('tests/conf/sgm_python_parall.json')
 
         # Instantiate machine
         pandora_machine = PandoraMachine()
@@ -101,7 +101,7 @@ class TestPluginPythonParall(unittest.TestCase):
         Test pandora + plugin_libsgm, with negative disparities
 
         """
-        user_cfg = pandora.read_config_file('conf/sgm_python_parall.json')
+        user_cfg = pandora.read_config_file('tests/conf/sgm_python_parall.json')
 
         # Import pandora plugins
         pandora.import_plugin()
@@ -146,7 +146,7 @@ class TestPluginPythonParall(unittest.TestCase):
         Test pandora + plugin_libsgm, with positive disparities
 
         """
-        user_cfg = pandora.read_config_file('conf/sgm_python_parall.json')
+        user_cfg = pandora.read_config_file('tests/conf/sgm_python_parall.json')
 
         # Import pandora plugins
         pandora.import_plugin()
@@ -182,7 +182,7 @@ class TestPluginPythonParall(unittest.TestCase):
         """
 
         # Prepare the configuration
-        user_cfg = pandora.read_config_file('conf/sgm_zncc_python_parall.json')
+        user_cfg = pandora.read_config_file('tests/conf/sgm_zncc_python_parall.json')
 
         # Import pandora plugins
         pandora.import_plugin()